backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Redis & Celery
redis>=5.0,<6.0
celery>=5.3,<6.0
django-celery-beat>=2.5,<3.0
django-celery-results>=2.5,<3.0

# OpenAI
openai>=1.6,<2.0

# Utilities
python-decouple>=3.8,<4.0
python-dotenv>=1.0,<2.0
Pillow>=10.1,<11.0
whitenoise>=6.6,<7.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Nested Routers
drf-nested-routers>=0.93,<1.0

# Development
black>=23.12,<24.0
flake8>=6.1,<7.0
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
